+++ image processing/code/Resize.cpp
@@ -2,10 +2,10 @@
 #include "Scale.h"
 
 template<qlm::ImageFormat frmt, qlm::pixel_t T>
-qlm::Image<frmt, T> qlm::Resize(const qlm::Image<frmt, T>& in, const unsigned int width, const unsigned int height, const qlm::InterpolationFlag method, const BorderMode<frmt, T>& border_mode)
+qlm::Image<frmt, T> qlm::Resize(const qlm::Image<frmt, T>& in, const size_t width, const size_t height, const qlm::InterpolationFlag method, const BorderMode<frmt, T>& border_mode)
 {
-	unsigned int img_width = in.Width();
-	unsigned int img_height = in.Height();
+	size_t img_width = in.Width();
+	size_t img_height = in.Height();
 	// calculate the scale factors
 	float scale_x = static_cast<float>(width) / img_width;
 	float scale_y = static_cast<float>(height) / img_height;
@@ -16,15 +16,15 @@ qlm::Image<frmt, T> qlm::Resize(const qlm::Image<frmt, T>& in, const unsigned in
 // Explicit instantiation for RGB , uint8_t
 template qlm::Image<qlm::ImageFormat::RGB, uint8_t>
 qlm::Resize<qlm::ImageFormat::RGB, uint8_t>(const qlm::Image<qlm::ImageFormat::RGB, uint8_t>& ,
-											const unsigned int,
-											const unsigned int,
+											const size_t,
+											const size_t,
 											const qlm::InterpolationFlag,
 											const BorderMode<qlm::ImageFormat::RGB, uint8_t>&);
 // Explicit instantiation for GRAY , uint8_t
 template qlm::Image<qlm::ImageFormat::GRAY, uint8_t>
 qlm::Resize<qlm::ImageFormat::GRAY, uint8_t>(const qlm::Image<qlm::ImageFormat::GRAY, uint8_t>&,
-											 const unsigned int,
-											 const unsigned int,
+											 const size_t,
+											 const size_t,
 											 const qlm::InterpolationFlag,
 											 const BorderMode<qlm::ImageFormat::GRAY, uint8_t>&);
