+++ image processing/include/FAST.h
@@ -9,7 +9,7 @@ namespace qlm
 	template<pixel_t T>
 	std::vector<KeyPoint<int>> FAST(
 		const Image<ImageFormat::GRAY, T>& in,
-		const unsigned int arc_length = 9,
+		const size_t arc_length = 9,
 		const T threshold = 10,
 		const bool non_max_suppression = true);
 }
